requirements.txt
alembic==0.9.2
appdirs==1.4.3
asn1crypto==0.22.0
beautifulsoup4==4.6.0
builtwith==1.3.2
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.3
click==6.7
cssselect==1.0.1
decorator==4.0.11
enum34==1.1.6
Flask==0.12.2
Flask-Admin==1.5.0
Flask-Migrate==2.0.3
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
ForgeryPy==0.1
idna==2.5
ipaddress==1.0.18
itsdangerous==0.24
Jinja2==2.9.6
lxml==3.7.3
Mako==1.0.6
MarkupSafe==1.0
packaging==16.8
parsel==1.2.0
pbr==3.0.1
pyasn1==0.2.3
pycparser==2.17
PyMySQL==0.7.11
pyparsing==2.2.0
python-dateutil==2.6.0
python-editor==1.0.3
queuelib==1.4.2
requests==2.17.3
six==1.10.0
SQLAlchemy==1.1.10
sqlalchemy-migrate==0.11.0
sqlparse==0.2.3
Tempita==0.5.2
urllib3==1.21.1
w3lib==1.17.0
Werkzeug==0.12.2
WTForms==2.1
